// connector-ui/src/app/core/services/models/nav-item-group.ts
/*
 * Copyright sovity GmbH and/or licensed to sovity GmbH under one or
 * more contributor license agreements. You may not use this file except
 * in compliance with the "Elastic License 2.0".
 *
 * SPDX-License-Identifier: Elastic-2.0
 */
import {NavItem} from './nav-item';

export interface NavItemGroup {
  title?: string;
  items: NavItem[];
}